+++ changelog.md
@@ -14,6 +14,9 @@
 - Add support for simple filter in the cli. Now `nagra select user -W
   '(=id 42)'` can now be written `nagra select user id=42`. See also
   the new guide for the command line in <examples/cli.md>
+- New methods in `Schema`: `get_table` and `get_view` explicitly
+  return a table and view, respectively. This avois the type ambiguity
+  of the `get` method.
 
 **Misc**:
 - Bug fix: When default value is provided for a column, it was ignored